# shared/openapi/openapi.json
{
  "components" : {
    "schemas" : {
      "Authentication" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "tokens" : {
            "$ref" : "#/components/schemas/Tokens"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AuthenticationForm" : {
        "properties" : {
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 12,
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ],
        "type" : "object"
      },
...
      "Link" : {
        "properties" : {
          "deprecation" : {
            "type" : "string"
          },
          "href" : {
            "type" : "string"
          },
          "hreflang" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Links" : {
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        },
        "type" : "object"
      },
...
      "RefreshForm" : {
        "properties" : {
          "refreshToken" : {
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
      "RegistrationForm" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 12,
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$",
            "type" : "string"
          },
          "passwordConfirmation" : {
            "maxLength" : 2147483647,
            "minLength" : 12,
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "password", "passwordConfirmation", "username" ],
        "type" : "object"
      },
      "RequestError" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "details" : {
            "type" : "object"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Tokens" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/auth/" : {
      "get" : {
        "operationId" : "isAuthenticated",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authentication"
                }
              }
            },
            "description" : "User is authenticated"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
    "/auth/check/{feature}" : {
      "post" : {
        "operationId" : "checkFeatureAccess",
        "parameters" : [ {
          "in" : "path",
          "name" : "feature",
          "required" : true,
          "schema" : {
            "enum" : [ "ADMINISTRATION", "CHALLENGE_CREATION" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "User is authorised"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
    "/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthenticationForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authentication"
                }
              }
            },
            "description" : "User authenticated successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
    "/auth/logout" : {
      "delete" : {
        "operationId" : "logout",
        "responses" : {
          "200" : {
            "description" : "User authenticated successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
    "/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authentication"
                }
              }
            },
            "description" : "Access token refreshed successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
    "/auth/register" : {
      "post" : {
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegistrationForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authentication"
                }
              }
            },
            "description" : "User authenticated successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "authentication-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}